requirements.txt
dlib
matplotlib
numpy
Pillow
Pillow
scikit_learn
scipy
torch
torch
torchvision
torchvision
tqdm
glob